Refactoring tests to match new code
